+++ test/oli/delivery/submission_test.exs
@@ -7,7 +7,7 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
   alias Oli.Delivery.Attempts.ActivityLifecycle
   alias Oli.Delivery.Attempts.ActivityLifecycle.Evaluate
 
-  alias Oli.Delivery.Attempts.Core.{ActivityAttempt, PartAttempt, StudentInput}
+  alias Oli.Delivery.Attempts.Core.{ResourceAccess, ActivityAttempt, PartAttempt, StudentInput}
   alias Oli.Delivery.Snapshots.Snapshot
   alias Oli.Delivery.Page.PageContext
   alias Oli.Delivery.Student.Summary
@@ -591,6 +591,43 @@ defmodule Oli.Delivery.AttemptsSubmissionTest do
       assert snapshot.activity_id == updated_attempt.resource_id
       assert snapshot.resource_id == page_revision.resource_id
       assert snapshot.revision_id == activity_revision.id
+    end
+
+    test "processing a submission where there are invalid scores", %{
+      graded_page_user1_attempt1: resource_attempt1,
+      user1_part1_attempt1: part_attempt,
+      section: section,
+      user1_activity_attempt1: activity_attempt
+    } do
+      part_inputs = [%{attempt_guid: part_attempt.attempt_guid, input: %StudentInput{input: "a"}}]
+
+      {:ok, _} =
+        Evaluate.evaluate_from_input(section.slug, activity_attempt.attempt_guid, part_inputs)
+
+      # verify the part attempt record was updated correctly
+      updated_attempt = Oli.Repo.get!(PartAttempt, part_attempt.id)
+      assert updated_attempt.score == 10
+      assert updated_attempt.out_of == 10
+      refute updated_attempt.date_evaluated == nil
+
+      # verify that the submission rolled up to the activity attempt
+      updated_attempt = Oli.Repo.get!(ActivityAttempt, activity_attempt.id)
+      assert updated_attempt.score == 1.0
+      assert updated_attempt.out_of == 1.0
+      refute updated_attempt.date_evaluated == nil
+
+      # Now simulate something having gone wrong, perhaps a rogue activity using
+      # client-side eval, or some other future bug in the system leading to an invalid
+      # score for the activity attempt
+
+      Attempts.update_activity_attempt(updated_attempt, %{score: 2.0, out_of: 1.0})
+
+      PageLifecycle.finalize(section.slug, resource_attempt1.attempt_guid)
+
+      # Verify a valid grade was recorded despite the invalid grade at the activity attempt
+      ra = Oli.Repo.get(ResourceAccess, resource_attempt1.resource_access_id)
+      assert ra.score == 1.0
+      assert ra.out_of == 1.0
     end
 
     test "handling reset attempts that request preservation of last attempt state", %{